lambda/requirements.txt
boto3==1.34.15
rawpy==0.24.0
imageio==2.33.0
pillow==10.2.0